refactore fem verification builder and minor fem element compatibility checks
