+++ pyproject.toml
@@ -57,7 +57,7 @@ ignore = ["E741"]
 
 [tool.isort]
 profile = "black"
-skip = [".pixi/", "node_modules/"]
+skip = [".pixi", "node_modules"]
 
 [tool.pixi.project]
 channels = ["https://repo.prefix.dev/code-aster", "conda-forge"]
@@ -68,7 +68,6 @@ prod = {features = ["py312"], solve-group = "prod"}
 lint = { features = ["lint"], no-default-feature = true}
 docs = { features = ["py312", "docs"], solve-group = "prod"}
 tests = { features = ["py312", "tests"], solve-group = "prod"}
-femtests = { features = ["py312", "tests", "femtests"], solve-group = "prod"}
 frontend = { features = ["web"], no-default-feature = true}
 
 [tool.pixi.feature.py312.dependencies]
@@ -82,9 +81,8 @@ ada-py = {path = ".", editable = true}
 
 [tool.pixi.feature.tests.dependencies]
 pytest = "*"
-
-[tool.pixi.feature.femtests.dependencies]
 paradoc = "*"
+python-dotenv = "*"
 
 [tool.pixi.feature.lint.dependencies]
 black = "24.4.2"
@@ -105,8 +103,7 @@ lint = "isort . && black . --config pyproject.toml && ruff check ."
 [tool.pixi.feature.docs.tasks]
 docs = "cd docs && make html"
 [tool.pixi.feature.tests.tasks]
-tests = "pytest tests"
-[tool.pixi.feature.femtests.tasks]
-ftests = { cmd="python build_verification_report.py", cwd="./tests/fem" }
+test = "pytest tests"
+ftest = { cmd="python build_verification_report.py", cwd="./tests/fem" }
 [tool.pixi.feature.web.tasks]
 npm = { cmd="npm", cwd="./src/frontend" }
