+++ tests/fem/.cache/cantilever_EIG_aba_line_o2_hqFalse_riFalse.json
@@ -1,0 +1,233 @@
+{
+    "name": "cantilever_EIG_aba_line_o2_hqFalse_riFalse",
+    "fem_format": "abaqus",
+    "metadata": {
+        "geo": "line",
+        "elo": 2,
+        "hexquad": false,
+        "reduced_integration": false
+    },
+    "eigen_mode_data": {
+        "1": {
+            "no": 1,
+            "f_hz": 10.786,
+            "eigenvalue": 4592.79,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "2": {
+            "no": 2,
+            "f_hz": 12.9321,
+            "eigenvalue": 6602.36,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "3": {
+            "no": 3,
+            "f_hz": 32.3579,
+            "eigenvalue": 41335.1,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "4": {
+            "no": 4,
+            "f_hz": 50.557,
+            "eigenvalue": 100907.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "5": {
+            "no": 5,
+            "f_hz": 53.9298,
+            "eigenvalue": 114820.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "6": {
+            "no": 6,
+            "f_hz": 75.5019,
+            "eigenvalue": 225048.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "7": {
+            "no": 7,
+            "f_hz": 79.6837,
+            "eigenvalue": 250668.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "8": {
+            "no": 8,
+            "f_hz": 97.0744,
+            "eigenvalue": 372022.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "9": {
+            "no": 9,
+            "f_hz": 118.648,
+            "eigenvalue": 555748.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "10": {
+            "no": 10,
+            "f_hz": 140.222,
+            "eigenvalue": 776236.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "11": {
+            "no": 11,
+            "f_hz": 161.799,
+            "eigenvalue": 1033510.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        }
+    },
+    "last_modified": 1723206676.3702
+}
