+++ tests/fem/.cache/cantilever_EIG_aba_solid_o1_hqTrue_riTrue.json
@@ -1,0 +1,233 @@
+{
+    "name": "cantilever_EIG_aba_solid_o1_hqTrue_riTrue",
+    "fem_format": "abaqus",
+    "metadata": {
+        "geo": "solid",
+        "elo": 1,
+        "hexquad": true,
+        "reduced_integration": true
+    },
+    "eigen_mode_data": {
+        "1": {
+            "no": 1,
+            "f_hz": 12.5136,
+            "eigenvalue": 6181.97,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "2": {
+            "no": 2,
+            "f_hz": 34.1571,
+            "eigenvalue": 46059.8,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "3": {
+            "no": 3,
+            "f_hz": 49.5976,
+            "eigenvalue": 97113.9,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "4": {
+            "no": 4,
+            "f_hz": 55.2711,
+            "eigenvalue": 120603.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "5": {
+            "no": 5,
+            "f_hz": 64.3052,
+            "eigenvalue": 163250.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "6": {
+            "no": 6,
+            "f_hz": 74.8656,
+            "eigenvalue": 221271.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "7": {
+            "no": 7,
+            "f_hz": 90.0328,
+            "eigenvalue": 320009.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "8": {
+            "no": 8,
+            "f_hz": 124.578,
+            "eigenvalue": 612693.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "9": {
+            "no": 9,
+            "f_hz": 155.992,
+            "eigenvalue": 960647.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "10": {
+            "no": 10,
+            "f_hz": 159.952,
+            "eigenvalue": 1010040.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "11": {
+            "no": 11,
+            "f_hz": 162.557,
+            "eigenvalue": 1043210.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        }
+    },
+    "last_modified": 1723206676.359144
+}
