+++ tests/fem/.cache/cantilever_EIG_aba_solid_o2_hqFalse_riFalse.json
@@ -1,0 +1,233 @@
+{
+    "name": "cantilever_EIG_aba_solid_o2_hqFalse_riFalse",
+    "fem_format": "abaqus",
+    "metadata": {
+        "geo": "solid",
+        "elo": 2,
+        "hexquad": false,
+        "reduced_integration": false
+    },
+    "eigen_mode_data": {
+        "1": {
+            "no": 1,
+            "f_hz": 12.9537,
+            "eigenvalue": 6624.46,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "2": {
+            "no": 2,
+            "f_hz": 20.2945,
+            "eigenvalue": 16259.8,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "3": {
+            "no": 3,
+            "f_hz": 50.5401,
+            "eigenvalue": 100840.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "4": {
+            "no": 4,
+            "f_hz": 77.1278,
+            "eigenvalue": 234845.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "5": {
+            "no": 5,
+            "f_hz": 97.8297,
+            "eigenvalue": 377834.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "6": {
+            "no": 6,
+            "f_hz": 132.481,
+            "eigenvalue": 692898.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "7": {
+            "no": 7,
+            "f_hz": 153.634,
+            "eigenvalue": 931826.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "8": {
+            "no": 8,
+            "f_hz": 178.118,
+            "eigenvalue": 1252490.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "9": {
+            "no": 9,
+            "f_hz": 215.822,
+            "eigenvalue": 1838870.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "10": {
+            "no": 10,
+            "f_hz": 253.203,
+            "eigenvalue": 2531030.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "11": {
+            "no": 11,
+            "f_hz": 254.466,
+            "eigenvalue": 2556350.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        }
+    },
+    "last_modified": 1723206677.394096
+}
