+++ tests/fem/.cache/cantilever_EIG_aba_solid_o2_hqTrue_riFalse.json
@@ -1,0 +1,233 @@
+{
+    "name": "cantilever_EIG_aba_solid_o2_hqTrue_riFalse",
+    "fem_format": "abaqus",
+    "metadata": {
+        "geo": "solid",
+        "elo": 2,
+        "hexquad": true,
+        "reduced_integration": false
+    },
+    "eigen_mode_data": {
+        "1": {
+            "no": 1,
+            "f_hz": 12.9551,
+            "eigenvalue": 6625.81,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "2": {
+            "no": 2,
+            "f_hz": 20.1503,
+            "eigenvalue": 16029.6,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "3": {
+            "no": 3,
+            "f_hz": 50.5412,
+            "eigenvalue": 100844.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "4": {
+            "no": 4,
+            "f_hz": 76.8903,
+            "eigenvalue": 233401.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "5": {
+            "no": 5,
+            "f_hz": 97.5939,
+            "eigenvalue": 376015.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "6": {
+            "no": 6,
+            "f_hz": 131.365,
+            "eigenvalue": 681274.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "7": {
+            "no": 7,
+            "f_hz": 149.958,
+            "eigenvalue": 887766.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "8": {
+            "no": 8,
+            "f_hz": 173.925,
+            "eigenvalue": 1194220.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "9": {
+            "no": 9,
+            "f_hz": 205.292,
+            "eigenvalue": 1663800.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "10": {
+            "no": 10,
+            "f_hz": 251.347,
+            "eigenvalue": 2494060.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "11": {
+            "no": 11,
+            "f_hz": 252.917,
+            "eigenvalue": 2525320.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        }
+    },
+    "last_modified": 1723206677.398095
+}
