+++ tests/fem/.cache/cantilever_EIG_aba_solid_o2_hqTrue_riTrue.json
@@ -1,0 +1,233 @@
+{
+    "name": "cantilever_EIG_aba_solid_o2_hqTrue_riTrue",
+    "fem_format": "abaqus",
+    "metadata": {
+        "geo": "solid",
+        "elo": 2,
+        "hexquad": true,
+        "reduced_integration": true
+    },
+    "eigen_mode_data": {
+        "1": {
+            "no": 1,
+            "f_hz": 12.9427,
+            "eigenvalue": 6613.22,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "2": {
+            "no": 2,
+            "f_hz": 20.0288,
+            "eigenvalue": 15836.8,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "3": {
+            "no": 3,
+            "f_hz": 50.4977,
+            "eigenvalue": 100671.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "4": {
+            "no": 4,
+            "f_hz": 76.6875,
+            "eigenvalue": 232171.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "5": {
+            "no": 5,
+            "f_hz": 97.3912,
+            "eigenvalue": 374455.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "6": {
+            "no": 6,
+            "f_hz": 130.809,
+            "eigenvalue": 675516.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "7": {
+            "no": 7,
+            "f_hz": 148.474,
+            "eigenvalue": 870287.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "8": {
+            "no": 8,
+            "f_hz": 172.171,
+            "eigenvalue": 1170260.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "9": {
+            "no": 9,
+            "f_hz": 201.839,
+            "eigenvalue": 1608310.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "10": {
+            "no": 10,
+            "f_hz": 246.165,
+            "eigenvalue": 2392280.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        },
+        "11": {
+            "no": 11,
+            "f_hz": 252.573,
+            "eigenvalue": 2518460.0,
+            "f_rad": null,
+            "f_imag_rad": null,
+            "px": null,
+            "py": null,
+            "pz": null,
+            "prx": null,
+            "pry": null,
+            "prz": null,
+            "efx": null,
+            "efy": null,
+            "efz": null,
+            "efrx": null,
+            "efry": null,
+            "efrz": null,
+            "source_dict": null
+        }
+    },
+    "last_modified": 1723206677.402096
+}
