--- tests/fem/.cache/cantilever_EIG_abaqus_LINE_o1_hqFalse.json
@@ -1,232 +1,0 @@
-{
-    "name": "cantilever_EIG_abaqus_LINE_o1_hqFalse",
-    "fem_format": "abaqus",
-    "metadata": {
-        "geo": "line",
-        "elo": 1,
-        "hexquad": false
-    },
-    "eigen_mode_data": {
-        "1": {
-            "no": 1,
-            "f_hz": 10.786,
-            "eigenvalue": 4592.5,
-            "f_rad": 67.768,
-            "f_imag_rad": null,
-            "px": 1.46422e-16,
-            "py": 6.26918e-15,
-            "pz": -1.20016e-15,
-            "prx": 1.2732,
-            "pry": 2.33525e-15,
-            "prz": 1.28888e-14,
-            "efx": 1.46422e-16,
-            "efy": 6.26918e-15,
-            "efz": -1.20016e-15,
-            "efrx": 1.2732,
-            "efry": 2.33525e-15,
-            "efrz": 1.28888e-14,
-            "source_dict": null
-        },
-        "2": {
-            "no": 2,
-            "f_hz": 12.93,
-            "eigenvalue": 6600.3,
-            "f_rad": 81.243,
-            "f_imag_rad": null,
-            "px": 1.86986e-17,
-            "py": 1.5639,
-            "pz": -2.94034e-16,
-            "prx": -0.78195,
-            "pry": 6.36856e-16,
-            "prz": 3.4088,
-            "efx": 1.86986e-17,
-            "efy": 1.5639,
-            "efz": -2.94034e-16,
-            "efrx": -0.78195,
-            "efry": 6.36856e-16,
-            "efrz": 3.4088,
-            "source_dict": null
-        },
-        "3": {
-            "no": 3,
-            "f_hz": 32.35,
-            "eigenvalue": 41314.0,
-            "f_rad": 203.26,
-            "f_imag_rad": null,
-            "px": -1.1623e-17,
-            "py": -4.05232e-16,
-            "pz": -9.13305e-16,
-            "prx": -0.42419,
-            "pry": 1.8723e-15,
-            "prz": -3.40967e-16,
-            "efx": -1.1623e-17,
-            "efy": -4.05232e-16,
-            "efz": -9.13305e-16,
-            "efrx": -0.42419,
-            "efry": 1.8723e-15,
-            "efrz": -3.40967e-16,
-            "source_dict": null
-        },
-        "4": {
-            "no": 4,
-            "f_hz": 50.55,
-            "eigenvalue": 100879.0,
-            "f_rad": 317.61,
-            "f_imag_rad": null,
-            "px": -4.00962e-17,
-            "py": -4.36392e-16,
-            "pz": 1.5385,
-            "prx": 0.76925,
-            "pry": -3.3527,
-            "prz": -8.7766e-17,
-            "efx": -4.00962e-17,
-            "efy": -4.36392e-16,
-            "efz": 1.5385,
-            "efrx": 0.76925,
-            "efry": -3.3527,
-            "efrz": -8.7766e-17,
-            "source_dict": null
-        },
-        "5": {
-            "no": 5,
-            "f_hz": 53.891,
-            "eigenvalue": 114656.0,
-            "f_rad": 338.61,
-            "f_imag_rad": null,
-            "px": 1.69091e-16,
-            "py": -2.96015e-15,
-            "pz": 4.21881e-14,
-            "prx": 0.25428,
-            "pry": -9.04286e-14,
-            "prz": -3.5694e-15,
-            "efx": 1.69091e-16,
-            "efy": -2.96015e-15,
-            "efz": 4.21881e-14,
-            "efrx": 0.25428,
-            "efry": -9.04286e-14,
-            "efrz": -3.5694e-15,
-            "source_dict": null
-        },
-        "6": {
-            "no": 6,
-            "f_hz": 75.396,
-            "eigenvalue": 224418.0,
-            "f_rad": 473.73,
-            "f_imag_rad": null,
-            "px": -4.56831e-16,
-            "py": 2.61623e-15,
-            "pz": -1.62207e-15,
-            "prx": -0.18138,
-            "pry": 3.02182e-15,
-            "prz": 1.98211e-15,
-            "efx": -4.56831e-16,
-            "efy": 2.61623e-15,
-            "efz": -1.62207e-15,
-            "efrx": -0.18138,
-            "efry": 3.02182e-15,
-            "efrz": 1.98211e-15,
-            "source_dict": null
-        },
-        "7": {
-            "no": 7,
-            "f_hz": 79.635,
-            "eigenvalue": 250364.0,
-            "f_rad": 500.36,
-            "f_imag_rad": null,
-            "px": 5.52771e-16,
-            "py": -0.86194,
-            "pz": 2.73012e-16,
-            "prx": 0.43097,
-            "pry": -2.54515e-16,
-            "prz": -0.53666,
-            "efx": 5.52771e-16,
-            "efy": -0.86194,
-            "efz": 2.73012e-16,
-            "efrx": 0.43097,
-            "efry": -2.54515e-16,
-            "efrz": -0.53666,
-            "source_dict": null
-        },
-        "8": {
-            "no": 8,
-            "f_hz": 96.849,
-            "eigenvalue": 370298.0,
-            "f_rad": 608.52,
-            "f_imag_rad": null,
-            "px": 1.71977e-15,
-            "py": 3.07376e-14,
-            "pz": -1.83359e-15,
-            "prx": -0.14082,
-            "pry": 4.10296e-15,
-            "prz": 2.02266e-15,
-            "efx": 1.71977e-15,
-            "efy": 3.07376e-14,
-            "efz": -1.83359e-15,
-            "efrx": -0.14082,
-            "efry": 4.10296e-15,
-            "efrz": 2.02266e-15,
-            "source_dict": null
-        },
-        "9": {
-            "no": 9,
-            "f_hz": 118.24,
-            "eigenvalue": 551898.0,
-            "f_rad": 742.9,
-            "f_imag_rad": null,
-            "px": 7.04262e-13,
-            "py": -4.90022e-13,
-            "pz": 4.60755e-13,
-            "prx": -0.11495,
-            "pry": 1.60122e-13,
-            "prz": -4.54911e-13,
-            "efx": 7.04262e-13,
-            "efy": -4.90022e-13,
-            "efz": 4.60755e-13,
-            "efrx": -0.11495,
-            "efry": 1.60122e-13,
-            "efrz": -4.54911e-13,
-            "source_dict": null
-        },
-        "10": {
-            "no": 10,
-            "f_hz": 139.54,
-            "eigenvalue": 768719.0,
-            "f_rad": 876.77,
-            "f_imag_rad": null,
-            "px": 2.30855e-11,
-            "py": 2.27309e-11,
-            "pz": 6.33182e-13,
-            "prx": 0.0969943,
-            "pry": 1.16022e-11,
-            "prz": -7.91023e-12,
-            "efx": 2.30855e-11,
-            "efy": 2.27309e-11,
-            "efz": 6.33182e-13,
-            "efrx": 0.0969943,
-            "efry": 1.16022e-11,
-            "efrz": -7.91023e-12,
-            "source_dict": null
-        },
-        "11": {
-            "no": 11,
-            "f_hz": 160.75,
-            "eigenvalue": 1020170.0,
-            "f_rad": 1010.0,
-            "f_imag_rad": null,
-            "px": 2.25619e-11,
-            "py": -3.58403e-10,
-            "pz": 1.26713e-10,
-            "prx": 0.083789,
-            "pry": -5.22137e-11,
-            "prz": -7.88836e-11,
-            "efx": 2.25619e-11,
-            "efy": -3.58403e-10,
-            "efz": 1.26713e-10,
-            "efrx": 0.083789,
-            "efry": -5.22137e-11,
-            "efrz": -7.88836e-11,
-            "source_dict": null
-        }
-    },
-    "last_modified": 1637068066.410009
-}
