--- tests/fem/.cache/cantilever_EIG_abaqus_LINE_o2_hqFalse.json
@@ -1,232 +1,0 @@
-{
-    "name": "cantilever_EIG_abaqus_LINE_o2_hqFalse",
-    "fem_format": "abaqus",
-    "metadata": {
-        "geo": "line",
-        "elo": 2,
-        "hexquad": false
-    },
-    "eigen_mode_data": {
-        "1": {
-            "no": 1,
-            "f_hz": 10.786,
-            "eigenvalue": 4592.8,
-            "f_rad": 67.77,
-            "f_imag_rad": null,
-            "px": 9.57168e-17,
-            "py": -1.97584e-16,
-            "pz": 1.28763e-17,
-            "prx": 1.2732,
-            "pry": 4.18341e-16,
-            "prz": -1.61755e-15,
-            "efx": 9.57168e-17,
-            "efy": -1.97584e-16,
-            "efz": 1.28763e-17,
-            "efrx": 1.2732,
-            "efry": 4.18341e-16,
-            "efrz": -1.61755e-15,
-            "source_dict": null
-        },
-        "2": {
-            "no": 2,
-            "f_hz": 12.932,
-            "eigenvalue": 6602.4,
-            "f_rad": 81.255,
-            "f_imag_rad": null,
-            "px": 2.36807e-17,
-            "py": 1.5642,
-            "pz": 6.7867e-17,
-            "prx": -0.78208,
-            "pry": -1.61105e-16,
-            "prz": 3.4089,
-            "efx": 2.36807e-17,
-            "efy": 1.5642,
-            "efz": 6.7867e-17,
-            "efrx": -0.78208,
-            "efry": -1.61105e-16,
-            "efrz": 3.4089,
-            "source_dict": null
-        },
-        "3": {
-            "no": 3,
-            "f_hz": 32.358,
-            "eigenvalue": 41335.0,
-            "f_rad": 203.31,
-            "f_imag_rad": null,
-            "px": -1.99304e-16,
-            "py": 3.8917e-15,
-            "pz": 5.68797e-15,
-            "prx": -0.42441,
-            "pry": -1.23869e-14,
-            "prz": 5.80482e-15,
-            "efx": -1.99304e-16,
-            "efy": 3.8917e-15,
-            "efz": 5.68797e-15,
-            "efrx": -0.42441,
-            "efry": -1.23869e-14,
-            "efrz": 5.80482e-15,
-            "source_dict": null
-        },
-        "4": {
-            "no": 4,
-            "f_hz": 50.557,
-            "eigenvalue": 100907.0,
-            "f_rad": 317.66,
-            "f_imag_rad": null,
-            "px": -3.26802e-16,
-            "py": 1.65357e-16,
-            "pz": 1.5387,
-            "prx": 0.76936,
-            "pry": -3.3528,
-            "prz": 1.35658e-16,
-            "efx": -3.26802e-16,
-            "efy": 1.65357e-16,
-            "efz": 1.5387,
-            "efrx": 0.76936,
-            "efry": -3.3528,
-            "efrz": 1.35658e-16,
-            "source_dict": null
-        },
-        "5": {
-            "no": 5,
-            "f_hz": 53.93,
-            "eigenvalue": 114820.0,
-            "f_rad": 338.85,
-            "f_imag_rad": null,
-            "px": 1.32508e-16,
-            "py": -1.16528e-15,
-            "pz": 1.14978e-14,
-            "prx": 0.25465,
-            "pry": -2.82393e-14,
-            "prz": -3.85076e-16,
-            "efx": 1.32508e-16,
-            "efy": -1.16528e-15,
-            "efz": 1.14978e-14,
-            "efrx": 0.25465,
-            "efry": -2.82393e-14,
-            "efrz": -3.85076e-16,
-            "source_dict": null
-        },
-        "6": {
-            "no": 6,
-            "f_hz": 75.502,
-            "eigenvalue": 225047.0,
-            "f_rad": 474.39,
-            "f_imag_rad": null,
-            "px": 5.23697e-16,
-            "py": -4.84333e-14,
-            "pz": 2.01797e-15,
-            "prx": -0.18189,
-            "pry": -1.31356e-14,
-            "prz": -3.15269e-14,
-            "efx": 5.23697e-16,
-            "efy": -4.84333e-14,
-            "efz": 2.01797e-15,
-            "efrx": -0.18189,
-            "efry": -1.31356e-14,
-            "efrz": -3.15269e-14,
-            "source_dict": null
-        },
-        "7": {
-            "no": 7,
-            "f_hz": 79.684,
-            "eigenvalue": 250668.0,
-            "f_rad": 500.67,
-            "f_imag_rad": null,
-            "px": -3.34262e-16,
-            "py": -0.86257,
-            "pz": 9.16429e-16,
-            "prx": 0.43129,
-            "pry": -1.54373e-15,
-            "prz": -0.537,
-            "efx": -3.34262e-16,
-            "efy": -0.86257,
-            "efz": 9.16429e-16,
-            "efrx": 0.43129,
-            "efry": -1.54373e-15,
-            "efrz": -0.537,
-            "source_dict": null
-        },
-        "8": {
-            "no": 8,
-            "f_hz": 97.074,
-            "eigenvalue": 372018.0,
-            "f_rad": 609.93,
-            "f_imag_rad": null,
-            "px": -3.62301e-15,
-            "py": -5.68982e-15,
-            "pz": 9.76116e-15,
-            "prx": -0.14147,
-            "pry": -5.51261e-15,
-            "prz": 8.85172e-16,
-            "efx": -3.62301e-15,
-            "efy": -5.68982e-15,
-            "efz": 9.76116e-15,
-            "efrx": -0.14147,
-            "efry": -5.51261e-15,
-            "efrz": 8.85172e-16,
-            "source_dict": null
-        },
-        "9": {
-            "no": 9,
-            "f_hz": 118.65,
-            "eigenvalue": 555733.0,
-            "f_rad": 745.48,
-            "f_imag_rad": null,
-            "px": 7.87292e-14,
-            "py": 1.15687e-13,
-            "pz": -1.12794e-13,
-            "prx": -0.11575,
-            "pry": 1.38245e-13,
-            "prz": -9.8574e-15,
-            "efx": 7.87292e-14,
-            "efy": 1.15687e-13,
-            "efz": -1.12794e-13,
-            "efrx": -0.11575,
-            "efry": 1.38245e-13,
-            "efrz": -9.8574e-15,
-            "source_dict": null
-        },
-        "10": {
-            "no": 10,
-            "f_hz": 140.22,
-            "eigenvalue": 776196.0,
-            "f_rad": 881.02,
-            "f_imag_rad": null,
-            "px": -2.28837e-11,
-            "py": 8.89721e-12,
-            "pz": 1.76998e-11,
-            "prx": 0.0979451,
-            "pry": -1.03007e-11,
-            "prz": 1.36529e-11,
-            "efx": -2.28837e-11,
-            "efy": 8.89721e-12,
-            "efz": 1.76998e-11,
-            "efrx": 0.0979451,
-            "efry": -1.03007e-11,
-            "efrz": 1.36529e-11,
-            "source_dict": null
-        },
-        "11": {
-            "no": 11,
-            "f_hz": 161.79,
-            "eigenvalue": 1033410.0,
-            "f_rad": 1016.6,
-            "f_imag_rad": null,
-            "px": -3.74279e-11,
-            "py": 1.75542e-11,
-            "pz": 4.37399e-11,
-            "prx": -0.0848882,
-            "pry": -2.79583e-11,
-            "prz": 2.10848e-11,
-            "efx": -3.74279e-11,
-            "efy": 1.75542e-11,
-            "efz": 4.37399e-11,
-            "efrx": -0.0848882,
-            "efry": -2.79583e-11,
-            "efrz": 2.10848e-11,
-            "source_dict": null
-        }
-    },
-    "last_modified": 1637068256.227119
-}
