--- tests/fem/.cache/cantilever_EIG_abaqus_SHELL_o1_hqFalse.json
@@ -1,232 +1,0 @@
-{
-    "name": "cantilever_EIG_abaqus_SHELL_o1_hqFalse",
-    "fem_format": "abaqus",
-    "metadata": {
-        "geo": "shell",
-        "elo": 1,
-        "hexquad": false
-    },
-    "eigen_mode_data": {
-        "1": {
-            "no": 1,
-            "f_hz": 13.561,
-            "eigenvalue": 7260.6,
-            "f_rad": 85.209,
-            "f_imag_rad": null,
-            "px": 1.4117e-05,
-            "py": 1.5788,
-            "pz": -4.24276e-05,
-            "prx": -0.78999,
-            "pry": 9.89872e-06,
-            "prz": 3.4406,
-            "efx": 1.4117e-05,
-            "efy": 1.5788,
-            "efz": -4.24276e-05,
-            "efrx": -0.78999,
-            "efry": 9.89872e-06,
-            "efrz": 3.4406,
-            "source_dict": null
-        },
-        "2": {
-            "no": 2,
-            "f_hz": 20.279,
-            "eigenvalue": 16235.0,
-            "f_rad": 127.42,
-            "f_imag_rad": null,
-            "px": 0.000182511,
-            "py": 0.00098669,
-            "pz": -4.21341e-05,
-            "prx": 0.30902,
-            "pry": 0.000141403,
-            "prz": 0.00106196,
-            "efx": 0.000182511,
-            "efy": 0.00098669,
-            "efz": -4.21341e-05,
-            "efrx": 0.30902,
-            "efry": 0.000141403,
-            "efrz": 0.00106196,
-            "source_dict": null
-        },
-        "3": {
-            "no": 3,
-            "f_hz": 52.11,
-            "eigenvalue": 107203.0,
-            "f_rad": 327.42,
-            "f_imag_rad": null,
-            "px": -0.000347948,
-            "py": 1.33172e-05,
-            "pz": 1.5412,
-            "prx": 0.77057,
-            "pry": -3.3545,
-            "prz": 0.000185973,
-            "efx": -0.000347948,
-            "efy": 1.33172e-05,
-            "efz": 1.5412,
-            "efrx": 0.77057,
-            "efry": -3.3545,
-            "efrz": 0.000185973,
-            "source_dict": null
-        },
-        "4": {
-            "no": 4,
-            "f_hz": 79.351,
-            "eigenvalue": 248582.0,
-            "f_rad": 498.58,
-            "f_imag_rad": null,
-            "px": 4.91514e-06,
-            "py": -0.9882,
-            "pz": -0.000228749,
-            "prx": 0.49455,
-            "pry": 8.60497e-05,
-            "prz": -0.61047,
-            "efx": 4.91514e-06,
-            "efy": -0.9882,
-            "efz": -0.000228749,
-            "efrx": 0.49455,
-            "efry": 8.60497e-05,
-            "efrz": -0.61047,
-            "source_dict": null
-        },
-        "5": {
-            "no": 5,
-            "f_hz": 101.28,
-            "eigenvalue": 404995.0,
-            "f_rad": 636.39,
-            "f_imag_rad": null,
-            "px": 0.000308331,
-            "py": -0.00259619,
-            "pz": 1.6777e-05,
-            "prx": -0.16764,
-            "pry": 8.15923e-05,
-            "prz": -0.000906037,
-            "efx": 0.000308331,
-            "efy": -0.00259619,
-            "efz": 1.6777e-05,
-            "efrx": -0.16764,
-            "efry": 8.15923e-05,
-            "efrz": -0.000906037,
-            "source_dict": null
-        },
-        "6": {
-            "no": 6,
-            "f_hz": 123.68,
-            "eigenvalue": 603906.0,
-            "f_rad": 777.11,
-            "f_imag_rad": null,
-            "px": -1.16403e-05,
-            "py": -0.0559252,
-            "pz": -0.00141037,
-            "prx": 0.0275745,
-            "pry": 0.00255713,
-            "prz": -0.069401,
-            "efx": -1.16403e-05,
-            "efy": -0.0559252,
-            "efz": -0.00141037,
-            "efrx": 0.0275745,
-            "efry": 0.00255713,
-            "efrz": -0.069401,
-            "source_dict": null
-        },
-        "7": {
-            "no": 7,
-            "f_hz": 144.35,
-            "eigenvalue": 822591.0,
-            "f_rad": 906.97,
-            "f_imag_rad": null,
-            "px": 1.13634e-05,
-            "py": 0.31733,
-            "pz": 0.00253778,
-            "prx": -0.1575,
-            "pry": -0.00165601,
-            "prz": 0.0699135,
-            "efx": 1.13634e-05,
-            "efy": 0.31733,
-            "efz": 0.00253778,
-            "efrx": -0.1575,
-            "efry": -0.00165601,
-            "efrz": 0.0699135,
-            "source_dict": null
-        },
-        "8": {
-            "no": 8,
-            "f_hz": 167.28,
-            "eigenvalue": 1104740.0,
-            "f_rad": 1051.1,
-            "f_imag_rad": null,
-            "px": 2.93983e-05,
-            "py": 0.55769,
-            "pz": -0.0050759,
-            "prx": -0.28178,
-            "pry": 0.00354952,
-            "prz": 0.24783,
-            "efx": 2.93983e-05,
-            "efy": 0.55769,
-            "efz": -0.0050759,
-            "efrx": -0.28178,
-            "efry": 0.00354952,
-            "efrz": 0.24783,
-            "source_dict": null
-        },
-        "9": {
-            "no": 9,
-            "f_hz": 196.32,
-            "eigenvalue": 1521600.0,
-            "f_rad": 1233.5,
-            "f_imag_rad": null,
-            "px": -2.85488e-05,
-            "py": -0.18105,
-            "pz": -0.00144469,
-            "prx": 0.0892653,
-            "pry": 0.00103899,
-            "prz": -0.0105054,
-            "efx": -2.85488e-05,
-            "efy": -0.18105,
-            "efz": -0.00144469,
-            "efrx": 0.0892653,
-            "efry": 0.00103899,
-            "efrz": -0.0105054,
-            "source_dict": null
-        },
-        "10": {
-            "no": 10,
-            "f_hz": 237.83,
-            "eigenvalue": 2232980.0,
-            "f_rad": 1494.3,
-            "f_imag_rad": null,
-            "px": -4.26287e-05,
-            "py": 0.3113,
-            "pz": -0.013715,
-            "prx": -0.16154,
-            "pry": 0.0077055,
-            "prz": 0.0708079,
-            "efx": -4.26287e-05,
-            "efy": 0.3113,
-            "efz": -0.013715,
-            "efrx": -0.16154,
-            "efry": 0.0077055,
-            "efrz": 0.0708079,
-            "source_dict": null
-        },
-        "11": {
-            "no": 11,
-            "f_hz": 257.19,
-            "eigenvalue": 2611350.0,
-            "f_rad": 1616.0,
-            "f_imag_rad": null,
-            "px": -0.000990975,
-            "py": -0.00280819,
-            "pz": -0.83447,
-            "prx": -0.41608,
-            "pry": 0.454,
-            "prz": -0.000428513,
-            "efx": -0.000990975,
-            "efy": -0.00280819,
-            "efz": -0.83447,
-            "efrx": -0.41608,
-            "efry": 0.454,
-            "efrz": -0.000428513,
-            "source_dict": null
-        }
-    },
-    "last_modified": 1637068121.926862
-}
