--- tests/fem/.cache/cantilever_EIG_abaqus_SHELL_o1_hqTrue.json
@@ -1,232 +1,0 @@
-{
-    "name": "cantilever_EIG_abaqus_SHELL_o1_hqTrue",
-    "fem_format": "abaqus",
-    "metadata": {
-        "geo": "shell",
-        "elo": 1,
-        "hexquad": true
-    },
-    "eigen_mode_data": {
-        "1": {
-            "no": 1,
-            "f_hz": 12.366,
-            "eigenvalue": 6036.8,
-            "f_rad": 77.697,
-            "f_imag_rad": null,
-            "px": 1.06958e-14,
-            "py": 1.5766,
-            "pz": -1.00408e-13,
-            "prx": -0.78831,
-            "pry": 2.29637e-13,
-            "prz": 3.4357,
-            "efx": 1.06958e-14,
-            "efy": 1.5766,
-            "efz": -1.00408e-13,
-            "efrx": -0.78831,
-            "efry": 2.29637e-13,
-            "efrz": 3.4357,
-            "source_dict": null
-        },
-        "2": {
-            "no": 2,
-            "f_hz": 19.215,
-            "eigenvalue": 14575.0,
-            "f_rad": 120.73,
-            "f_imag_rad": null,
-            "px": -1.42272e-14,
-            "py": 1.06966e-11,
-            "pz": 6.39062e-13,
-            "prx": 0.30824,
-            "pry": -1.40947e-12,
-            "prz": 2.33298e-11,
-            "efx": -1.42272e-14,
-            "efy": 1.06966e-11,
-            "efz": 6.39062e-13,
-            "efrx": 0.30824,
-            "efry": -1.40947e-12,
-            "efrz": 2.33298e-11,
-            "source_dict": null
-        },
-        "3": {
-            "no": 3,
-            "f_hz": 51.903,
-            "eigenvalue": 106353.0,
-            "f_rad": 326.12,
-            "f_imag_rad": null,
-            "px": 2.21662e-15,
-            "py": 6.07158e-14,
-            "pz": 1.5393,
-            "prx": 0.76965,
-            "pry": -3.3502,
-            "prz": 1.91339e-13,
-            "efx": 2.21662e-15,
-            "efy": 6.07158e-14,
-            "efz": 1.5393,
-            "efrx": 0.76965,
-            "efry": -3.3502,
-            "efrz": 1.91339e-13,
-            "source_dict": null
-        },
-        "4": {
-            "no": 4,
-            "f_hz": 73.396,
-            "eigenvalue": 212671.0,
-            "f_rad": 461.16,
-            "f_imag_rad": null,
-            "px": -1.81689e-15,
-            "py": -0.98101,
-            "pz": -8.20046e-14,
-            "prx": 0.49051,
-            "pry": 1.69644e-13,
-            "prz": -0.60754,
-            "efx": -1.81689e-15,
-            "efy": -0.98101,
-            "efz": -8.20046e-14,
-            "efrx": 0.49051,
-            "efry": 1.69644e-13,
-            "efrz": -0.60754,
-            "source_dict": null
-        },
-        "5": {
-            "no": 5,
-            "f_hz": 93.615,
-            "eigenvalue": 345980.0,
-            "f_rad": 588.2,
-            "f_imag_rad": null,
-            "px": 1.76146e-14,
-            "py": -5.62157e-13,
-            "pz": -4.67124e-14,
-            "prx": -0.16765,
-            "pry": 8.69581e-14,
-            "prz": -3.16447e-13,
-            "efx": 1.76146e-14,
-            "efy": -5.62157e-13,
-            "efz": -4.67124e-14,
-            "efrx": -0.16765,
-            "efry": 8.69581e-14,
-            "efrz": -3.16447e-13,
-            "source_dict": null
-        },
-        "6": {
-            "no": 6,
-            "f_hz": 126.06,
-            "eigenvalue": 627351.0,
-            "f_rad": 792.05,
-            "f_imag_rad": null,
-            "px": -8.8949e-15,
-            "py": -0.016235,
-            "pz": -7.74333e-14,
-            "prx": 0.00811752,
-            "pry": 1.57349e-13,
-            "prz": -0.0461725,
-            "efx": -8.8949e-15,
-            "efy": -0.016235,
-            "efz": -7.74333e-14,
-            "efrx": 0.00811752,
-            "efry": 1.57349e-13,
-            "efrz": -0.0461725,
-            "source_dict": null
-        },
-        "7": {
-            "no": 7,
-            "f_hz": 143.41,
-            "eigenvalue": 811931.0,
-            "f_rad": 901.07,
-            "f_imag_rad": null,
-            "px": 2.27803e-15,
-            "py": 0.4207,
-            "pz": -3.57138e-14,
-            "prx": -0.21035,
-            "pry": 1.17147e-14,
-            "prz": 0.11912,
-            "efx": 2.27803e-15,
-            "efy": 0.4207,
-            "efz": -3.57138e-14,
-            "efrx": -0.21035,
-            "efry": 1.17147e-14,
-            "efrz": 0.11912,
-            "source_dict": null
-        },
-        "8": {
-            "no": 8,
-            "f_hz": 167.17,
-            "eigenvalue": 1103300.0,
-            "f_rad": 1050.4,
-            "f_imag_rad": null,
-            "px": 1.10267e-12,
-            "py": 0.55136,
-            "pz": -1.24278e-12,
-            "prx": -0.27568,
-            "pry": 9.1448e-13,
-            "prz": 0.24242,
-            "efx": 1.10267e-12,
-            "efy": 0.55136,
-            "efz": -1.24278e-12,
-            "efrx": -0.27568,
-            "efry": 9.1448e-13,
-            "efrz": 0.24242,
-            "source_dict": null
-        },
-        "9": {
-            "no": 9,
-            "f_hz": 195.19,
-            "eigenvalue": 1504050.0,
-            "f_rad": 1226.4,
-            "f_imag_rad": null,
-            "px": -6.15778e-13,
-            "py": -0.18997,
-            "pz": 3.49686e-12,
-            "prx": 0.0949873,
-            "pry": -1.21219e-12,
-            "prz": -0.0145471,
-            "efx": -6.15778e-13,
-            "efy": -0.18997,
-            "efz": 3.49686e-12,
-            "efrx": 0.0949873,
-            "efry": -1.21219e-12,
-            "efrz": -0.0145471,
-            "source_dict": null
-        },
-        "10": {
-            "no": 10,
-            "f_hz": 237.78,
-            "eigenvalue": 2232120.0,
-            "f_rad": 1494.0,
-            "f_imag_rad": null,
-            "px": -5.85587e-11,
-            "py": 0.32799,
-            "pz": 7.42599e-10,
-            "prx": -0.16399,
-            "pry": -2.8644e-10,
-            "prz": 0.0748145,
-            "efx": -5.85587e-11,
-            "efy": 0.32799,
-            "efz": 7.42599e-10,
-            "efrx": -0.16399,
-            "efry": -2.8644e-10,
-            "efrz": 0.0748145,
-            "source_dict": null
-        },
-        "11": {
-            "no": 11,
-            "f_hz": 244.56,
-            "eigenvalue": 2361150.0,
-            "f_rad": 1536.6,
-            "f_imag_rad": null,
-            "px": 3.35608e-10,
-            "py": 7.07286e-10,
-            "pz": 2.55693e-10,
-            "prx": 0.0993088,
-            "pry": 3.39666e-11,
-            "prz": -1.60886e-10,
-            "efx": 3.35608e-10,
-            "efy": 7.07286e-10,
-            "efz": 2.55693e-10,
-            "efrx": 0.0993088,
-            "efry": 3.39666e-11,
-            "efrz": -1.60886e-10,
-            "source_dict": null
-        }
-    },
-    "last_modified": 1637068148.239984
-}
