--- tests/fem/.cache/cantilever_EIG_abaqus_SHELL_o2_hqFalse.json
@@ -1,232 +1,0 @@
-{
-    "name": "cantilever_EIG_abaqus_SHELL_o2_hqFalse",
-    "fem_format": "abaqus",
-    "metadata": {
-        "geo": "shell",
-        "elo": 2,
-        "hexquad": false
-    },
-    "eigen_mode_data": {
-        "1": {
-            "no": 1,
-            "f_hz": 12.755,
-            "eigenvalue": 6422.4,
-            "f_rad": 80.14,
-            "f_imag_rad": null,
-            "px": 5.85185e-07,
-            "py": 1.5774,
-            "pz": -1.35815e-06,
-            "prx": -0.78867,
-            "pry": 3.46292e-06,
-            "prz": 3.4376,
-            "efx": 5.85185e-07,
-            "efy": 1.5774,
-            "efz": -1.35815e-06,
-            "efrx": -0.78867,
-            "efry": 3.46292e-06,
-            "efrz": 3.4376,
-            "source_dict": null
-        },
-        "2": {
-            "no": 2,
-            "f_hz": 19.564,
-            "eigenvalue": 15110.0,
-            "f_rad": 122.92,
-            "f_imag_rad": null,
-            "px": 2.23863e-07,
-            "py": 3.30368e-05,
-            "pz": -1.09734e-05,
-            "prx": -0.30845,
-            "pry": 2.24235e-05,
-            "prz": 8.17744e-05,
-            "efx": 2.23863e-07,
-            "efy": 3.30368e-05,
-            "efz": -1.09734e-05,
-            "efrx": -0.30845,
-            "efry": 2.24235e-05,
-            "efrz": 8.17744e-05,
-            "source_dict": null
-        },
-        "3": {
-            "no": 3,
-            "f_hz": 51.967,
-            "eigenvalue": 106614.0,
-            "f_rad": 326.52,
-            "f_imag_rad": null,
-            "px": 2.23372e-06,
-            "py": -6.4576e-06,
-            "pz": 1.5392,
-            "prx": 0.76959,
-            "pry": -3.35,
-            "prz": -1.90742e-06,
-            "efx": 2.23372e-06,
-            "efy": -6.4576e-06,
-            "efz": 1.5392,
-            "efrx": 0.76959,
-            "efry": -3.35,
-            "efrz": -1.90742e-06,
-            "source_dict": null
-        },
-        "4": {
-            "no": 4,
-            "f_hz": 75.145,
-            "eigenvalue": 222925.0,
-            "f_rad": 472.15,
-            "f_imag_rad": null,
-            "px": -4.19237e-06,
-            "py": -0.98588,
-            "pz": -8.77986e-06,
-            "prx": 0.49293,
-            "pry": 2.05576e-05,
-            "prz": -0.60988,
-            "efx": -4.19237e-06,
-            "efy": -0.98588,
-            "efz": -8.77986e-06,
-            "efrx": 0.49293,
-            "efry": 2.05576e-05,
-            "efrz": -0.60988,
-            "source_dict": null
-        },
-        "5": {
-            "no": 5,
-            "f_hz": 95.992,
-            "eigenvalue": 363774.0,
-            "f_rad": 603.14,
-            "f_imag_rad": null,
-            "px": -1.84941e-06,
-            "py": -4.47882e-05,
-            "pz": -1.71311e-05,
-            "prx": 0.16763,
-            "pry": 4.65701e-05,
-            "prz": -4.02217e-05,
-            "efx": -1.84941e-06,
-            "efy": -4.47882e-05,
-            "efz": -1.71311e-05,
-            "efrx": 0.16763,
-            "efry": 4.65701e-05,
-            "efrz": -4.02217e-05,
-            "source_dict": null
-        },
-        "6": {
-            "no": 6,
-            "f_hz": 125.22,
-            "eigenvalue": 619068.0,
-            "f_rad": 786.81,
-            "f_imag_rad": null,
-            "px": -1.76192e-06,
-            "py": -0.0269103,
-            "pz": -7.87073e-06,
-            "prx": 0.0134504,
-            "pry": -1.58741e-07,
-            "prz": -0.053885,
-            "efx": -1.76192e-06,
-            "efy": -0.0269103,
-            "efz": -7.87073e-06,
-            "efrx": 0.0134504,
-            "efry": -1.58741e-07,
-            "efrz": -0.053885,
-            "source_dict": null
-        },
-        "7": {
-            "no": 7,
-            "f_hz": 142.81,
-            "eigenvalue": 805177.0,
-            "f_rad": 897.32,
-            "f_imag_rad": null,
-            "px": 8.10995e-07,
-            "py": 0.40449,
-            "pz": -9.0366e-06,
-            "prx": -0.20225,
-            "pry": -8.04501e-06,
-            "prz": 0.10963,
-            "efx": 8.10995e-07,
-            "efy": 0.40449,
-            "efz": -9.0366e-06,
-            "efrx": -0.20225,
-            "efry": -8.04501e-06,
-            "efrz": 0.10963,
-            "source_dict": null
-        },
-        "8": {
-            "no": 8,
-            "f_hz": 166.62,
-            "eigenvalue": 1096020.0,
-            "f_rad": 1046.9,
-            "f_imag_rad": null,
-            "px": 1.00916e-05,
-            "py": 0.54171,
-            "pz": 1.53306e-05,
-            "prx": -0.27084,
-            "pry": -1.19317e-05,
-            "prz": 0.24125,
-            "efx": 1.00916e-05,
-            "efy": 0.54171,
-            "efz": 1.53306e-05,
-            "efrx": -0.27084,
-            "efry": -1.19317e-05,
-            "efrz": 0.24125,
-            "source_dict": null
-        },
-        "9": {
-            "no": 9,
-            "f_hz": 192.96,
-            "eigenvalue": 1469890.0,
-            "f_rad": 1212.4,
-            "f_imag_rad": null,
-            "px": 3.67064e-06,
-            "py": -0.19577,
-            "pz": 3.56656e-05,
-            "prx": 0.0979023,
-            "pry": -6.1257e-06,
-            "prz": -0.0153434,
-            "efx": 3.67064e-06,
-            "efy": -0.19577,
-            "efz": 3.56656e-05,
-            "efrx": 0.0979023,
-            "efry": -6.1257e-06,
-            "efrz": -0.0153434,
-            "source_dict": null
-        },
-        "10": {
-            "no": 10,
-            "f_hz": 233.59,
-            "eigenvalue": 2154050.0,
-            "f_rad": 1467.7,
-            "f_imag_rad": null,
-            "px": 1.36562e-06,
-            "py": 0.31452,
-            "pz": -1.74713e-05,
-            "prx": -0.15726,
-            "pry": -5.28922e-06,
-            "prz": 0.069908,
-            "efx": 1.36562e-06,
-            "efy": 0.31452,
-            "efz": -1.74713e-05,
-            "efrx": -0.15726,
-            "efry": -5.28922e-06,
-            "efrz": 0.069908,
-            "source_dict": null
-        },
-        "11": {
-            "no": 11,
-            "f_hz": 250.12,
-            "eigenvalue": 2469710.0,
-            "f_rad": 1571.5,
-            "f_imag_rad": null,
-            "px": -8.17968e-06,
-            "py": 5.67043e-05,
-            "pz": 0.000290466,
-            "prx": 0.0984288,
-            "pry": -0.00015004,
-            "prz": 1.66444e-05,
-            "efx": -8.17968e-06,
-            "efy": 5.67043e-05,
-            "efz": 0.000290466,
-            "efrx": 0.0984288,
-            "efry": -0.00015004,
-            "efrz": 1.66444e-05,
-            "source_dict": null
-        }
-    },
-    "last_modified": 1637068364.542264
-}
