--- tests/fem/.cache/cantilever_EIG_abaqus_SHELL_o2_hqTrue.json
@@ -1,232 +1,0 @@
-{
-    "name": "cantilever_EIG_abaqus_SHELL_o2_hqTrue",
-    "fem_format": "abaqus",
-    "metadata": {
-        "geo": "shell",
-        "elo": 2,
-        "hexquad": true
-    },
-    "eigen_mode_data": {
-        "1": {
-            "no": 1,
-            "f_hz": 12.757,
-            "eigenvalue": 6425.2,
-            "f_rad": 80.158,
-            "f_imag_rad": null,
-            "px": 2.77776e-14,
-            "py": 1.5774,
-            "pz": 2.20973e-13,
-            "prx": -0.7887,
-            "pry": -4.80544e-13,
-            "prz": 3.4379,
-            "efx": 2.77776e-14,
-            "efy": 1.5774,
-            "efz": 2.20973e-13,
-            "efrx": -0.7887,
-            "efry": -4.80544e-13,
-            "efrz": 3.4379,
-            "source_dict": null
-        },
-        "2": {
-            "no": 2,
-            "f_hz": 19.485,
-            "eigenvalue": 14988.0,
-            "f_rad": 122.43,
-            "f_imag_rad": null,
-            "px": 3.16726e-14,
-            "py": 1.94511e-10,
-            "pz": -8.90443e-13,
-            "prx": 0.30847,
-            "pry": 1.92136e-12,
-            "prz": 4.24273e-10,
-            "efx": 3.16726e-14,
-            "efy": 1.94511e-10,
-            "efz": -8.90443e-13,
-            "efrx": 0.30847,
-            "efry": 1.92136e-12,
-            "efrz": 4.24273e-10,
-            "source_dict": null
-        },
-        "3": {
-            "no": 3,
-            "f_hz": 51.994,
-            "eigenvalue": 106726.0,
-            "f_rad": 326.69,
-            "f_imag_rad": null,
-            "px": -1.28866e-14,
-            "py": -7.15687e-14,
-            "pz": 1.5392,
-            "prx": 0.76958,
-            "pry": -3.3506,
-            "prz": -3.7341e-13,
-            "efx": -1.28866e-14,
-            "efy": -7.15687e-14,
-            "efz": 1.5392,
-            "efrx": 0.76958,
-            "efry": -3.3506,
-            "efrz": -3.7341e-13,
-            "source_dict": null
-        },
-        "4": {
-            "no": 4,
-            "f_hz": 75.143,
-            "eigenvalue": 222911.0,
-            "f_rad": 472.13,
-            "f_imag_rad": null,
-            "px": -1.10611e-14,
-            "py": -0.9874,
-            "pz": 3.80736e-13,
-            "prx": 0.4937,
-            "pry": -8.34353e-13,
-            "prz": -0.61114,
-            "efx": -1.10611e-14,
-            "efy": -0.9874,
-            "efz": 3.80736e-13,
-            "efrx": 0.4937,
-            "efry": -8.34353e-13,
-            "efrz": -0.61114,
-            "source_dict": null
-        },
-        "5": {
-            "no": 5,
-            "f_hz": 95.911,
-            "eigenvalue": 363156.0,
-            "f_rad": 602.62,
-            "f_imag_rad": null,
-            "px": -8.0028e-15,
-            "py": -5.00772e-12,
-            "pz": -8.28623e-14,
-            "prx": -0.16766,
-            "pry": 1.87906e-13,
-            "prz": -3.20758e-12,
-            "efx": -8.0028e-15,
-            "efy": -5.00772e-12,
-            "efz": -8.28623e-14,
-            "efrx": -0.16766,
-            "efry": 1.87906e-13,
-            "efrz": -3.20758e-12,
-            "source_dict": null
-        },
-        "6": {
-            "no": 6,
-            "f_hz": 124.99,
-            "eigenvalue": 616786.0,
-            "f_rad": 785.36,
-            "f_imag_rad": null,
-            "px": -2.48756e-14,
-            "py": -0.0269998,
-            "pz": 4.20371e-13,
-            "prx": 0.0134999,
-            "pry": -7.80831e-13,
-            "prz": -0.0539768,
-            "efx": -2.48756e-14,
-            "efy": -0.0269998,
-            "efz": 4.20371e-13,
-            "efrx": 0.0134999,
-            "efry": -7.80831e-13,
-            "efrz": -0.0539768,
-            "source_dict": null
-        },
-        "7": {
-            "no": 7,
-            "f_hz": 142.31,
-            "eigenvalue": 799551.0,
-            "f_rad": 894.18,
-            "f_imag_rad": null,
-            "px": 2.77583e-15,
-            "py": 0.39938,
-            "pz": -6.91601e-14,
-            "prx": -0.19969,
-            "pry": 1.16889e-13,
-            "prz": 0.10753,
-            "efx": 2.77583e-15,
-            "efy": 0.39938,
-            "efz": -6.91601e-14,
-            "efrx": -0.19969,
-            "efry": 1.16889e-13,
-            "efrz": 0.10753,
-            "source_dict": null
-        },
-        "8": {
-            "no": 8,
-            "f_hz": 165.83,
-            "eigenvalue": 1085650.0,
-            "f_rad": 1041.9,
-            "f_imag_rad": null,
-            "px": 1.02904e-14,
-            "py": 0.541,
-            "pz": 4.94607e-13,
-            "prx": -0.2705,
-            "pry": -9.86543e-13,
-            "prz": 0.24142,
-            "efx": 1.02904e-14,
-            "efy": 0.541,
-            "efz": 4.94607e-13,
-            "efrx": -0.2705,
-            "efry": -9.86543e-13,
-            "efrz": 0.24142,
-            "source_dict": null
-        },
-        "9": {
-            "no": 9,
-            "f_hz": 192.16,
-            "eigenvalue": 1457690.0,
-            "f_rad": 1207.3,
-            "f_imag_rad": null,
-            "px": -1.06694e-13,
-            "py": -0.19409,
-            "pz": 7.57107e-13,
-            "prx": 0.0970433,
-            "pry": -4.57432e-13,
-            "prz": -0.0148533,
-            "efx": -1.06694e-13,
-            "efy": -0.19409,
-            "efz": 7.57107e-13,
-            "efrx": 0.0970433,
-            "efry": -4.57432e-13,
-            "efrz": -0.0148533,
-            "source_dict": null
-        },
-        "10": {
-            "no": 10,
-            "f_hz": 232.82,
-            "eigenvalue": 2139850.0,
-            "f_rad": 1462.8,
-            "f_imag_rad": null,
-            "px": -2.88659e-11,
-            "py": 0.31373,
-            "pz": 2.18783e-10,
-            "prx": -0.15686,
-            "pry": -6.35185e-11,
-            "prz": 0.0698999,
-            "efx": -2.88659e-11,
-            "efy": 0.31373,
-            "efz": 2.18783e-10,
-            "efrx": -0.15686,
-            "efry": -6.35185e-11,
-            "efrz": 0.0698999,
-            "source_dict": null
-        },
-        "11": {
-            "no": 11,
-            "f_hz": 250.14,
-            "eigenvalue": 2470080.0,
-            "f_rad": 1571.7,
-            "f_imag_rad": null,
-            "px": 8.26946e-13,
-            "py": 1.2194e-10,
-            "pz": -2.59456e-11,
-            "prx": -0.0983521,
-            "pry": -3.14133e-12,
-            "prz": 2.58267e-11,
-            "efx": 8.26946e-13,
-            "efy": 1.2194e-10,
-            "efz": -2.59456e-11,
-            "efrx": -0.0983521,
-            "efry": -3.14133e-12,
-            "efrz": 2.58267e-11,
-            "source_dict": null
-        }
-    },
-    "last_modified": 1637068389.230473
-}
