--- tests/fem/.cache/cantilever_EIG_abaqus_SOLID_o1_hqFalse.json
@@ -1,232 +1,0 @@
-{
-    "name": "cantilever_EIG_abaqus_SOLID_o1_hqFalse",
-    "fem_format": "abaqus",
-    "metadata": {
-        "geo": "solid",
-        "elo": 1,
-        "hexquad": false
-    },
-    "eigen_mode_data": {
-        "1": {
-            "no": 1,
-            "f_hz": 14.33,
-            "eigenvalue": 8106.4,
-            "f_rad": 90.036,
-            "f_imag_rad": null,
-            "px": 5.08704e-06,
-            "py": 1.5691,
-            "pz": -0.00110131,
-            "prx": -0.78374,
-            "pry": 0.00229974,
-            "prz": 3.4233,
-            "efx": 5.08704e-06,
-            "efy": 1.5691,
-            "efz": -0.00110131,
-            "efrx": -0.78374,
-            "efry": 0.00229974,
-            "efrz": 3.4233,
-            "source_dict": null
-        },
-        "2": {
-            "no": 2,
-            "f_hz": 50.977,
-            "eigenvalue": 102592.0,
-            "f_rad": 320.3,
-            "f_imag_rad": null,
-            "px": -7.65874e-06,
-            "py": 0.00045318,
-            "pz": 1.5396,
-            "prx": 0.76883,
-            "pry": -3.3547,
-            "prz": 0.00184644,
-            "efx": -7.65874e-06,
-            "efy": 0.00045318,
-            "efz": 1.5396,
-            "efrx": 0.76883,
-            "efry": -3.3547,
-            "efrz": 0.00184644,
-            "source_dict": null
-        },
-        "3": {
-            "no": 3,
-            "f_hz": 71.588,
-            "eigenvalue": 202323.0,
-            "f_rad": 449.8,
-            "f_imag_rad": null,
-            "px": -1.51926e-05,
-            "py": 0.0060931,
-            "pz": -0.00237862,
-            "prx": -0.2734,
-            "pry": 0.00628165,
-            "prz": 0.0120717,
-            "efx": -1.51926e-05,
-            "efy": 0.0060931,
-            "efz": -0.00237862,
-            "efrx": -0.2734,
-            "efry": 0.00628165,
-            "efrz": 0.0120717,
-            "source_dict": null
-        },
-        "4": {
-            "no": 4,
-            "f_hz": 87.846,
-            "eigenvalue": 304654.0,
-            "f_rad": 551.95,
-            "f_imag_rad": null,
-            "px": -9.62106e-05,
-            "py": -0.8634,
-            "pz": -0.00103763,
-            "prx": 0.43302,
-            "pry": 0.00249528,
-            "prz": -0.54337,
-            "efx": -9.62106e-05,
-            "efy": -0.8634,
-            "efz": -0.00103763,
-            "efrx": 0.43302,
-            "efry": 0.00249528,
-            "efrz": -0.54337,
-            "source_dict": null
-        },
-        "5": {
-            "no": 5,
-            "f_hz": 232.23,
-            "eigenvalue": 2129120.0,
-            "f_rad": 1459.1,
-            "f_imag_rad": null,
-            "px": 0.000490257,
-            "py": 0.0922779,
-            "pz": -0.0015269,
-            "prx": 0.0807324,
-            "pry": -0.00125464,
-            "prz": 0.0334743,
-            "efx": 0.000490257,
-            "efy": 0.0922779,
-            "efz": -0.0015269,
-            "efrx": 0.0807324,
-            "efry": -0.00125464,
-            "efrz": 0.0334743,
-            "source_dict": null
-        },
-        "6": {
-            "no": 6,
-            "f_hz": 239.99,
-            "eigenvalue": 2273790.0,
-            "f_rad": 1507.9,
-            "f_imag_rad": null,
-            "px": 0.000105469,
-            "py": 0.56679,
-            "pz": -0.0123151,
-            "prx": -0.30829,
-            "pry": 0.00768983,
-            "prz": 0.21481,
-            "efx": 0.000105469,
-            "efy": 0.56679,
-            "efz": -0.0123151,
-            "efrx": -0.30829,
-            "efry": 0.00768983,
-            "efrz": 0.21481,
-            "source_dict": null
-        },
-        "7": {
-            "no": 7,
-            "f_hz": 256.69,
-            "eigenvalue": 2601210.0,
-            "f_rad": 1612.8,
-            "f_imag_rad": null,
-            "px": -1.32083e-05,
-            "py": -0.00560498,
-            "pz": -0.80359,
-            "prx": -0.39937,
-            "pry": 0.44664,
-            "prz": -0.00221932,
-            "efx": -1.32083e-05,
-            "efy": -0.00560498,
-            "efz": -0.80359,
-            "efrx": -0.39937,
-            "efry": 0.44664,
-            "efrz": -0.00221932,
-            "source_dict": null
-        },
-        "8": {
-            "no": 8,
-            "f_hz": 425.51,
-            "eigenvalue": 7147840.0,
-            "f_rad": 2673.5,
-            "f_imag_rad": null,
-            "px": 0.0159749,
-            "py": -0.00416396,
-            "pz": 0.00300288,
-            "prx": -0.0834909,
-            "pry": 0.0083144,
-            "prz": -0.00871255,
-            "efx": 0.0159749,
-            "efy": -0.00416396,
-            "efz": 0.00300288,
-            "efrx": -0.0834909,
-            "efry": 0.0083144,
-            "efrz": -0.00871255,
-            "source_dict": null
-        },
-        "9": {
-            "no": 9,
-            "f_hz": 432.08,
-            "eigenvalue": 7370190.0,
-            "f_rad": 2714.8,
-            "f_imag_rad": null,
-            "px": 1.2747,
-            "py": 0.000201396,
-            "pz": 1.44778e-05,
-            "prx": 0.000265532,
-            "pry": 0.63735,
-            "prz": -0.63734,
-            "efx": 1.2747,
-            "efy": 0.000201396,
-            "efz": 1.44778e-05,
-            "efrx": 0.000265532,
-            "efry": 0.63735,
-            "efrz": -0.63734,
-            "source_dict": null
-        },
-        "10": {
-            "no": 10,
-            "f_hz": 452.01,
-            "eigenvalue": 8066130.0,
-            "f_rad": 2840.1,
-            "f_imag_rad": null,
-            "px": 0.00117057,
-            "py": -0.36688,
-            "pz": 0.00120726,
-            "prx": 0.18377,
-            "pry": 2.80348e-05,
-            "prz": -0.0979261,
-            "efx": 0.00117057,
-            "efy": -0.36688,
-            "efz": 0.00120726,
-            "efrx": 0.18377,
-            "efry": 2.80348e-05,
-            "efrz": -0.0979261,
-            "source_dict": null
-        },
-        "11": {
-            "no": 11,
-            "f_hz": 584.5,
-            "eigenvalue": 13487400.0,
-            "f_rad": 3672.5,
-            "f_imag_rad": null,
-            "px": -0.000106738,
-            "py": 0.000853148,
-            "pz": 0.42735,
-            "prx": 0.2135,
-            "pry": -0.12606,
-            "prz": 0.000321715,
-            "efx": -0.000106738,
-            "efy": 0.000853148,
-            "efz": 0.42735,
-            "efrx": 0.2135,
-            "efry": -0.12606,
-            "efrz": 0.000321715,
-            "source_dict": null
-        }
-    },
-    "last_modified": 1637068207.319815
-}
