--- tests/fem/.cache/cantilever_EIG_abaqus_SOLID_o1_hqTrue.json
@@ -1,232 +1,0 @@
-{
-    "name": "cantilever_EIG_abaqus_SOLID_o1_hqTrue",
-    "fem_format": "abaqus",
-    "metadata": {
-        "geo": "solid",
-        "elo": 1,
-        "hexquad": true
-    },
-    "eigen_mode_data": {
-        "1": {
-            "no": 1,
-            "f_hz": 13.165,
-            "eigenvalue": 6842.5,
-            "f_rad": 82.72,
-            "f_imag_rad": null,
-            "px": 1.00939e-13,
-            "py": 1.5649,
-            "pz": -2.55974e-13,
-            "prx": -0.78245,
-            "pry": 6.63465e-13,
-            "prz": 3.4128,
-            "efx": 1.00939e-13,
-            "efy": 1.5649,
-            "efz": -2.55974e-13,
-            "efrx": -0.78245,
-            "efry": 6.63465e-13,
-            "efrz": 3.4128,
-            "source_dict": null
-        },
-        "2": {
-            "no": 2,
-            "f_hz": 20.545,
-            "eigenvalue": 16664.0,
-            "f_rad": 129.09,
-            "f_imag_rad": null,
-            "px": 2.95698e-14,
-            "py": 1.8462e-10,
-            "pz": -1.19314e-11,
-            "prx": 0.30616,
-            "pry": 2.59374e-11,
-            "prz": 4.07995e-10,
-            "efx": 2.95698e-14,
-            "efy": 1.8462e-10,
-            "efz": -1.19314e-11,
-            "efrx": 0.30616,
-            "efry": 2.59374e-11,
-            "efrz": 4.07995e-10,
-            "source_dict": null
-        },
-        "3": {
-            "no": 3,
-            "f_hz": 50.611,
-            "eigenvalue": 101122.0,
-            "f_rad": 318.0,
-            "f_imag_rad": null,
-            "px": 5.93316e-14,
-            "py": 7.23565e-13,
-            "pz": 1.539,
-            "prx": 0.76948,
-            "pry": -3.3525,
-            "prz": 8.41876e-13,
-            "efx": 5.93316e-14,
-            "efy": 7.23565e-13,
-            "efz": 1.539,
-            "efrx": 0.76948,
-            "efry": -3.3525,
-            "efrz": 8.41876e-13,
-            "source_dict": null
-        },
-        "4": {
-            "no": 4,
-            "f_hz": 80.979,
-            "eigenvalue": 258882.0,
-            "f_rad": 508.8,
-            "f_imag_rad": null,
-            "px": -2.13871e-14,
-            "py": -0.8812,
-            "pz": 4.65691e-13,
-            "prx": 0.4406,
-            "pry": -1.4234e-12,
-            "prz": -0.55189,
-            "efx": -2.13871e-14,
-            "efy": -0.8812,
-            "efz": 4.65691e-13,
-            "efrx": 0.4406,
-            "efry": -1.4234e-12,
-            "efrz": -0.55189,
-            "source_dict": null
-        },
-        "5": {
-            "no": 5,
-            "f_hz": 98.936,
-            "eigenvalue": 386431.0,
-            "f_rad": 621.64,
-            "f_imag_rad": null,
-            "px": 5.22957e-14,
-            "py": -1.24665e-11,
-            "pz": 9.79207e-15,
-            "prx": -0.16486,
-            "pry": 4.72046e-13,
-            "prz": -1.41801e-11,
-            "efx": 5.22957e-14,
-            "efy": -1.24665e-11,
-            "efz": 9.79207e-15,
-            "efrx": -0.16486,
-            "efry": 4.72046e-13,
-            "efrz": -1.41801e-11,
-            "source_dict": null
-        },
-        "6": {
-            "no": 6,
-            "f_hz": 214.81,
-            "eigenvalue": 1821690.0,
-            "f_rad": 1349.7,
-            "f_imag_rad": null,
-            "px": -4.29445e-14,
-            "py": 0.57277,
-            "pz": 1.91914e-13,
-            "prx": -0.28639,
-            "pry": 3.0498e-13,
-            "prz": 0.21688,
-            "efx": -4.29445e-14,
-            "efy": 0.57277,
-            "efz": 1.91914e-13,
-            "efrx": -0.28639,
-            "efry": 3.0498e-13,
-            "efrz": 0.21688,
-            "source_dict": null
-        },
-        "7": {
-            "no": 7,
-            "f_hz": 254.46,
-            "eigenvalue": 2556170.0,
-            "f_rad": 1598.8,
-            "f_imag_rad": null,
-            "px": 7.8405e-14,
-            "py": 3.6512e-13,
-            "pz": -0.80982,
-            "prx": -0.40491,
-            "pry": 0.44835,
-            "prz": 2.20747e-13,
-            "efx": 7.8405e-14,
-            "efy": 3.6512e-13,
-            "efz": -0.80982,
-            "efrx": -0.40491,
-            "efry": 0.44835,
-            "efrz": 2.20747e-13,
-            "source_dict": null
-        },
-        "8": {
-            "no": 8,
-            "f_hz": 256.63,
-            "eigenvalue": 2600010.0,
-            "f_rad": 1612.5,
-            "f_imag_rad": null,
-            "px": 8.97713e-14,
-            "py": 2.70612e-12,
-            "pz": -3.5354e-12,
-            "prx": 0.098542,
-            "pry": 2.02877e-12,
-            "prz": 1.39311e-13,
-            "efx": 8.97713e-14,
-            "efy": 2.70612e-12,
-            "efz": -3.5354e-12,
-            "efrx": 0.098542,
-            "efry": 2.02877e-12,
-            "efrz": 1.39311e-13,
-            "source_dict": null
-        },
-        "9": {
-            "no": 9,
-            "f_hz": 354.41,
-            "eigenvalue": 4958660.0,
-            "f_rad": 2226.8,
-            "f_imag_rad": null,
-            "px": 1.86465e-11,
-            "py": 0.40337,
-            "pz": -5.94524e-12,
-            "prx": -0.20168,
-            "pry": 9.47672e-12,
-            "prz": 0.10581,
-            "efx": 1.86465e-11,
-            "efy": 0.40337,
-            "efz": -5.94524e-12,
-            "efrx": -0.20168,
-            "efry": 9.47672e-12,
-            "efrz": 0.10581,
-            "source_dict": null
-        },
-        "10": {
-            "no": 10,
-            "f_hz": 398.71,
-            "eigenvalue": 6275890.0,
-            "f_rad": 2505.2,
-            "f_imag_rad": null,
-            "px": -9.36167e-11,
-            "py": 0.0143574,
-            "pz": 2.74956e-11,
-            "prx": -0.00717868,
-            "pry": -4.81952e-11,
-            "prz": 0.00737971,
-            "efx": -9.36167e-11,
-            "efy": 0.0143574,
-            "efz": 2.74956e-11,
-            "efrx": -0.00717868,
-            "efry": -4.81952e-11,
-            "efrz": 0.00737971,
-            "source_dict": null
-        },
-        "11": {
-            "no": 11,
-            "f_hz": 410.14,
-            "eigenvalue": 6640880.0,
-            "f_rad": 2577.0,
-            "f_imag_rad": null,
-            "px": 1.66592e-10,
-            "py": -0.000988999,
-            "pz": -1.4914e-10,
-            "prx": 0.0004945,
-            "pry": 1.03061e-10,
-            "prz": -0.0116402,
-            "efx": 1.66592e-10,
-            "efy": -0.000988999,
-            "efz": -1.4914e-10,
-            "efrx": 0.0004945,
-            "efry": 1.03061e-10,
-            "efrz": -0.0116402,
-            "source_dict": null
-        }
-    },
-    "last_modified": 1637068231.945674
-}
