--- tests/fem/.cache/cantilever_EIG_abaqus_SOLID_o2_hqFalse.json
@@ -1,232 +1,0 @@
-{
-    "name": "cantilever_EIG_abaqus_SOLID_o2_hqFalse",
-    "fem_format": "abaqus",
-    "metadata": {
-        "geo": "solid",
-        "elo": 2,
-        "hexquad": false
-    },
-    "eigen_mode_data": {
-        "1": {
-            "no": 1,
-            "f_hz": 12.948,
-            "eigenvalue": 6618.1,
-            "f_rad": 81.352,
-            "f_imag_rad": null,
-            "px": -3.98141e-08,
-            "py": 1.5753,
-            "pz": 2.74687e-06,
-            "prx": -0.78764,
-            "pry": -5.77023e-06,
-            "prz": 3.4337,
-            "efx": -3.98141e-08,
-            "efy": 1.5753,
-            "efz": 2.74687e-06,
-            "efrx": -0.78764,
-            "efry": -5.77023e-06,
-            "efrz": 3.4337,
-            "source_dict": null
-        },
-        "2": {
-            "no": 2,
-            "f_hz": 20.183,
-            "eigenvalue": 16081.0,
-            "f_rad": 126.81,
-            "f_imag_rad": null,
-            "px": -1.91364e-07,
-            "py": 3.84843e-05,
-            "pz": -1.39532e-06,
-            "prx": -0.30695,
-            "pry": 1.9665e-06,
-            "prz": 0.000101579,
-            "efx": -1.91364e-07,
-            "efy": 3.84843e-05,
-            "efz": -1.39532e-06,
-            "efrx": -0.30695,
-            "efry": 1.9665e-06,
-            "efrz": 0.000101579,
-            "source_dict": null
-        },
-        "3": {
-            "no": 3,
-            "f_hz": 50.517,
-            "eigenvalue": 100747.0,
-            "f_rad": 317.41,
-            "f_imag_rad": null,
-            "px": -3.95233e-06,
-            "py": 1.60256e-05,
-            "pz": 1.5404,
-            "prx": 0.77019,
-            "pry": -3.3541,
-            "prz": 4.50012e-06,
-            "efx": -3.95233e-06,
-            "efy": 1.60256e-05,
-            "efz": 1.5404,
-            "efrx": 0.77019,
-            "efry": -3.3541,
-            "efrz": 4.50012e-06,
-            "source_dict": null
-        },
-        "4": {
-            "no": 4,
-            "f_hz": 76.898,
-            "eigenvalue": 233450.0,
-            "f_rad": 483.17,
-            "f_imag_rad": null,
-            "px": 3.43233e-07,
-            "py": -0.96467,
-            "pz": 2.69411e-05,
-            "prx": 0.48235,
-            "pry": -5.12918e-05,
-            "prz": -0.59841,
-            "efx": 3.43233e-07,
-            "efy": -0.96467,
-            "efz": 2.69411e-05,
-            "efrx": 0.48235,
-            "efry": -5.12918e-05,
-            "efrz": -0.59841,
-            "source_dict": null
-        },
-        "5": {
-            "no": 5,
-            "f_hz": 97.614,
-            "eigenvalue": 376172.0,
-            "f_rad": 613.33,
-            "f_imag_rad": null,
-            "px": 3.04286e-06,
-            "py": 3.99833e-06,
-            "pz": -3.07603e-06,
-            "prx": 0.1653,
-            "pry": 9.13255e-06,
-            "prz": 9.51775e-06,
-            "efx": 3.04286e-06,
-            "efy": 3.99833e-06,
-            "efz": -3.07603e-06,
-            "efrx": 0.1653,
-            "efry": 9.13255e-06,
-            "efrz": 9.51775e-06,
-            "source_dict": null
-        },
-        "6": {
-            "no": 6,
-            "f_hz": 131.69,
-            "eigenvalue": 684688.0,
-            "f_rad": 827.46,
-            "f_imag_rad": null,
-            "px": -2.30147e-06,
-            "py": -0.0206742,
-            "pz": -3.28912e-05,
-            "prx": 0.0103696,
-            "pry": 2.46683e-05,
-            "prz": -0.045002,
-            "efx": -2.30147e-06,
-            "efy": -0.0206742,
-            "efz": -3.28912e-05,
-            "efrx": 0.0103696,
-            "efry": 2.46683e-05,
-            "efrz": -0.045002,
-            "source_dict": null
-        },
-        "7": {
-            "no": 7,
-            "f_hz": 150.98,
-            "eigenvalue": 899899.0,
-            "f_rad": 948.63,
-            "f_imag_rad": null,
-            "px": -8.69952e-06,
-            "py": 0.37896,
-            "pz": -0.000124355,
-            "prx": -0.18951,
-            "pry": 7.53499e-05,
-            "prz": 0.10487,
-            "efx": -8.69952e-06,
-            "efy": 0.37896,
-            "efz": -0.000124355,
-            "efrx": -0.18951,
-            "efry": 7.53499e-05,
-            "efrz": 0.10487,
-            "source_dict": null
-        },
-        "8": {
-            "no": 8,
-            "f_hz": 175.18,
-            "eigenvalue": 1211550.0,
-            "f_rad": 1100.7,
-            "f_imag_rad": null,
-            "px": -3.40877e-07,
-            "py": 0.54182,
-            "pz": -6.27402e-05,
-            "prx": -0.2709,
-            "pry": 4.67992e-05,
-            "prz": 0.23468,
-            "efx": -3.40877e-07,
-            "efy": 0.54182,
-            "efz": -6.27402e-05,
-            "efrx": -0.2709,
-            "efry": 4.67992e-05,
-            "efrz": 0.23468,
-            "source_dict": null
-        },
-        "9": {
-            "no": 9,
-            "f_hz": 208.15,
-            "eigenvalue": 1710440.0,
-            "f_rad": 1307.8,
-            "f_imag_rad": null,
-            "px": 1.85533e-05,
-            "py": -0.15647,
-            "pz": 0.000191809,
-            "prx": 0.0782259,
-            "pry": -8.66662e-05,
-            "prz": -0.00738803,
-            "efx": 1.85533e-05,
-            "efy": -0.15647,
-            "efz": 0.000191809,
-            "efrx": 0.0782259,
-            "efry": -8.66662e-05,
-            "efrz": -0.00738803,
-            "source_dict": null
-        },
-        "10": {
-            "no": 10,
-            "f_hz": 252.85,
-            "eigenvalue": 2524010.0,
-            "f_rad": 1588.7,
-            "f_imag_rad": null,
-            "px": 1.97333e-05,
-            "py": 8.03134e-05,
-            "pz": -0.82406,
-            "prx": -0.41211,
-            "pry": 0.45416,
-            "prz": 8.2723e-06,
-            "efx": 1.97333e-05,
-            "efy": 8.03134e-05,
-            "efz": -0.82406,
-            "efrx": -0.41211,
-            "efry": 0.45416,
-            "efrz": 8.2723e-06,
-            "source_dict": null
-        },
-        "11": {
-            "no": 11,
-            "f_hz": 254.03,
-            "eigenvalue": 2547610.0,
-            "f_rad": 1596.1,
-            "f_imag_rad": null,
-            "px": -1.64792e-05,
-            "py": 0.00576473,
-            "pz": 0.000600843,
-            "prx": -0.0999719,
-            "pry": -0.000341117,
-            "prz": 0.00144265,
-            "efx": -1.64792e-05,
-            "efy": 0.00576473,
-            "efz": 0.000600843,
-            "efrx": -0.0999719,
-            "efry": -0.000341117,
-            "efrz": 0.00144265,
-            "source_dict": null
-        }
-    },
-    "last_modified": 1637068523.718023
-}
