--- tests/fem/.cache/cantilever_EIG_abaqus_SOLID_o2_hqTrue.json
@@ -1,232 +1,0 @@
-{
-    "name": "cantilever_EIG_abaqus_SOLID_o2_hqTrue",
-    "fem_format": "abaqus",
-    "metadata": {
-        "geo": "solid",
-        "elo": 2,
-        "hexquad": true
-    },
-    "eigen_mode_data": {
-        "1": {
-            "no": 1,
-            "f_hz": 12.942,
-            "eigenvalue": 6612.8,
-            "f_rad": 81.319,
-            "f_imag_rad": null,
-            "px": 1.13499e-13,
-            "py": 1.5759,
-            "pz": -2.60085e-08,
-            "prx": -0.78797,
-            "pry": 5.10845e-08,
-            "prz": 3.4349,
-            "efx": 1.13499e-13,
-            "efy": 1.5759,
-            "efz": -2.60085e-08,
-            "efrx": -0.78797,
-            "efry": 5.10845e-08,
-            "efrz": 3.4349,
-            "source_dict": null
-        },
-        "2": {
-            "no": 2,
-            "f_hz": 20.026,
-            "eigenvalue": 15832.0,
-            "f_rad": 125.83,
-            "f_imag_rad": null,
-            "px": -1.92158e-10,
-            "py": 1.73894e-09,
-            "pz": 8.9321e-12,
-            "prx": 0.30708,
-            "pry": -1.15046e-10,
-            "prz": 3.9418e-09,
-            "efx": -1.92158e-10,
-            "efy": 1.73894e-09,
-            "efz": 8.9321e-12,
-            "efrx": 0.30708,
-            "efry": -1.15046e-10,
-            "efrz": 3.9418e-09,
-            "source_dict": null
-        },
-        "3": {
-            "no": 3,
-            "f_hz": 50.495,
-            "eigenvalue": 100662.0,
-            "f_rad": 317.27,
-            "f_imag_rad": null,
-            "px": -2.4705e-13,
-            "py": -8.83724e-09,
-            "pz": 1.5406,
-            "prx": 0.77028,
-            "pry": -3.3542,
-            "prz": 3.65129e-08,
-            "efx": -2.4705e-13,
-            "efy": -8.83724e-09,
-            "efz": 1.5406,
-            "efrx": 0.77028,
-            "efry": -3.3542,
-            "efrz": 3.65129e-08,
-            "source_dict": null
-        },
-        "4": {
-            "no": 4,
-            "f_hz": 76.686,
-            "eigenvalue": 232163.0,
-            "f_rad": 481.83,
-            "f_imag_rad": null,
-            "px": -9.20066e-14,
-            "py": -0.97692,
-            "pz": 3.52206e-07,
-            "prx": 0.48846,
-            "pry": -3.62623e-08,
-            "prz": -0.6055,
-            "efx": -9.20066e-14,
-            "efy": -0.97692,
-            "efz": 3.52206e-07,
-            "efrx": 0.48846,
-            "efry": -3.62623e-08,
-            "efrz": -0.6055,
-            "source_dict": null
-        },
-        "5": {
-            "no": 5,
-            "f_hz": 97.385,
-            "eigenvalue": 374405.0,
-            "f_rad": 611.89,
-            "f_imag_rad": null,
-            "px": 1.66991e-10,
-            "py": -1.42801e-10,
-            "pz": 3.79118e-12,
-            "prx": -0.16539,
-            "pry": 7.42417e-11,
-            "prz": -2.32516e-10,
-            "efx": 1.66991e-10,
-            "efy": -1.42801e-10,
-            "efz": 3.79118e-12,
-            "efrx": -0.16539,
-            "efry": 7.42417e-11,
-            "efrz": -2.32516e-10,
-            "source_dict": null
-        },
-        "6": {
-            "no": 6,
-            "f_hz": 130.8,
-            "eigenvalue": 675416.0,
-            "f_rad": 821.84,
-            "f_imag_rad": null,
-            "px": -3.45501e-14,
-            "py": -0.0185329,
-            "pz": -1.34476e-07,
-            "prx": 0.00926636,
-            "pry": -1.56064e-06,
-            "prz": -0.0458406,
-            "efx": -3.45501e-14,
-            "efy": -0.0185329,
-            "efz": -1.34476e-07,
-            "efrx": 0.00926636,
-            "efry": -1.56064e-06,
-            "efrz": -0.0458406,
-            "source_dict": null
-        },
-        "7": {
-            "no": 7,
-            "f_hz": 148.46,
-            "eigenvalue": 870139.0,
-            "f_rad": 932.81,
-            "f_imag_rad": null,
-            "px": -8.46422e-15,
-            "py": 0.39375,
-            "pz": -4.10493e-06,
-            "prx": -0.19688,
-            "pry": 2.17803e-06,
-            "prz": 0.10894,
-            "efx": -8.46422e-15,
-            "efy": 0.39375,
-            "efz": -4.10493e-06,
-            "efrx": -0.19688,
-            "efry": 2.17803e-06,
-            "efrz": 0.10894,
-            "source_dict": null
-        },
-        "8": {
-            "no": 8,
-            "f_hz": 172.18,
-            "eigenvalue": 1170330.0,
-            "f_rad": 1081.8,
-            "f_imag_rad": null,
-            "px": -1.72822e-13,
-            "py": 0.53807,
-            "pz": 4.93359e-06,
-            "prx": -0.26903,
-            "pry": -3.03563e-06,
-            "prz": 0.23632,
-            "efx": -1.72822e-13,
-            "efy": 0.53807,
-            "efz": 4.93359e-06,
-            "efrx": -0.26903,
-            "efry": -3.03563e-06,
-            "efrz": 0.23632,
-            "source_dict": null
-        },
-        "9": {
-            "no": 9,
-            "f_hz": 201.81,
-            "eigenvalue": 1607850.0,
-            "f_rad": 1268.0,
-            "f_imag_rad": null,
-            "px": 4.45465e-14,
-            "py": -0.17444,
-            "pz": 1.89023e-06,
-            "prx": 0.0872192,
-            "pry": -6.45026e-07,
-            "prz": -0.0113095,
-            "efx": 4.45465e-14,
-            "efy": -0.17444,
-            "efz": 1.89023e-06,
-            "efrx": 0.0872192,
-            "efry": -6.45026e-07,
-            "efrz": -0.0113095,
-            "source_dict": null
-        },
-        "10": {
-            "no": 10,
-            "f_hz": 246.12,
-            "eigenvalue": 2391360.0,
-            "f_rad": 1546.4,
-            "f_imag_rad": null,
-            "px": 6.21819e-12,
-            "py": 0.31458,
-            "pz": 5.56096e-06,
-            "prx": -0.15729,
-            "pry": -3.27974e-06,
-            "prz": 0.0732133,
-            "efx": 6.21819e-12,
-            "efy": 0.31458,
-            "efz": 5.56096e-06,
-            "efrx": -0.15729,
-            "efry": -3.27974e-06,
-            "efrz": 0.0732133,
-            "source_dict": null
-        },
-        "11": {
-            "no": 11,
-            "f_hz": 252.54,
-            "eigenvalue": 2517820.0,
-            "f_rad": 1586.8,
-            "f_imag_rad": null,
-            "px": 4.85467e-14,
-            "py": 7.64256e-07,
-            "pz": -0.82623,
-            "prx": -0.41312,
-            "pry": 0.45496,
-            "prz": 2.80323e-07,
-            "efx": 4.85467e-14,
-            "efy": 7.64256e-07,
-            "efz": -0.82623,
-            "efrx": -0.41312,
-            "efry": 0.45496,
-            "efrz": 2.80323e-07,
-            "source_dict": null
-        }
-    },
-    "last_modified": 1637068553.421876
-}
