+++ tests/fem/.cache/software_versions.json
@@ -1,0 +1,6 @@
+{
+    "calculix": "2.21",
+    "abaqus": "2024 (RELr426 190762)",
+    "code_aster": "17.1.0",
+    "sesam": "10.17-02"
+}
